Add unit tests and Readme usage examples
