+++ pyrate_limiter/__init__.py
@@ -7,4 +7,4 @@ from .request_rate import *
 from .constants import *
 from .exceptions import *
 
-__version__ = "2.2.2"
+__version__ = "2.3.0"
