+++ setup.cfg
@@ -2,11 +2,16 @@
 based_on_style = pep8
 spaces_before_comment = 4
 
-
 [flake8]
 max-line-length = 120
 ignore = E401
 
+# Code Climate still uses the 'pep8' section; renamed to 'pycodestyle' when used locally
+[pep8]
+max-line-length = 120
+
+[pycodestyle]
+max-line-length = 120
 
 [tool.pylint]
 disable = C0330
